Using existing discovery.rb file instead of new discovery_Config file for discovery connection
