+++ lib/nexpose.rb
@@ -72,11 +72,11 @@ require 'nexpose/site_credentials'
 require 'nexpose/shared_credential'
 require 'nexpose/data_table'
 require 'nexpose/device'
-require 'nexpose/discovery'
-require 'nexpose/discovery/filter'
 require 'nexpose/engine'
 require 'nexpose/external'
 require 'nexpose/filter'
+require 'nexpose/discovery'
+require 'nexpose/discovery/filter'
 require 'nexpose/global_settings'
 require 'nexpose/group'
 require 'nexpose/dag'
@@ -101,7 +101,6 @@ require 'nexpose/vuln_exception'
 require 'nexpose/connection'
 require 'nexpose/maint'
 require 'nexpose/version'
-require 'nexpose/discovery_config'
 
 module Nexpose
